Add files via upload
Fine-tuning a sample of "Consumer complaints narrative"  and "Product" data in the complaint.csv data.
This is the second  part of the process which consist of fine-tuning the balanced sample of 60000 entries that has 10000 entries of each of the 'product' counts. The model was trained on 5 epochs and the learning rate was set at 1e-5 and batch size 32. Total training time taken across all epochs [Validation and model saving time included]:  3:16:30  (h:mm:ss). The best model is saved at epoch 4. The training macro f1-score: 0.89 .The validation macro f1-score: 0.85