Add files via upload
Fine-tuning a sample of "Consumer complaints narrative"  and "Product" data in the complaint.csv data.
This is the third part of the process which consist of predicting and evaluating the DistilBERT model performance on the test data.The test data consist of 6000 entries. The total time taken to create the data frame with top 3 rank of predictions in the test data is : 0:36:05 (h:mm:ss). The macro-averaged f1-score on the test data is 0.87