Add files via upload
This is the second part of the process which consist of fine-tuning DistilBERT on dataset_clean. The model was trained on 5 epochs and the learning rate was set at 2e-5 and batch size 16. Total training time taken across all epochs [Validation and model saving time included]: 0:01:26 (h:mm:ss). The best model is saved at epoch 5. The training macro f1-score: 0.26. The validation macro f1-score: 0.07