Add files via upload
Fine-tuning DistilBERT on RBI dataset.
This is the first part of the process which consist of removing certain stop-words, the newline character, parenthesis and the text present within parenthesis from the columns 'Principle' and 'Indian Position'. Duplicate entries and text with word count less than 7 are removed. The training data consist of the 'Topic' column and the preprocessed 'Principle' column. The validation and testing data consist of the 'Topic' column and the preprocessed 'Indian Position' column. The split data is saved as data_clean.