Add files via upload
Fine-tuning a sample of "Consumer complaints narrative"  and "Product" data in the complaint.csv data.
This is the first part of the process which consist of preprocessing the "Consumer complaints narrative" and creating a balanced sample of 60000 entries that has 10000 entries of each of the 'product' counts. The stop-words are removed using Gensim library.